REF: separate pcds config from compare utility, pull out constant paths for interactive use
